+++ cupy/_indexing/__init__.py
@@ -1,0 +1,2 @@
+# Functions from the following NumPy document
+# https://docs.scipy.org/doc/numpy/reference/routines.indexing.html
