+++ cupy/_indexing/generate.py
