+++ cupy/_indexing/iterate.py
