--- cupy/indexing/__init__.py
@@ -1,8 +1,0 @@
-# Functions from the following NumPy document
-# https://docs.scipy.org/doc/numpy/reference/routines.indexing.html
-
-# "NOQA" to suppress flake8 warning
-from cupy.indexing import generate  # NOQA
-from cupy.indexing import indexing  # NOQA
-from cupy.indexing import insert  # NOQA
-from cupy.indexing import iterate  # NOQA
